refactor(components): [TreeSelect] move utility functions to utils.ts
